% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariatenormal.R
\name{dmvnorm}
\alias{dmvnorm}
\title{Evaluate density of multivariate Normal}
\usage{
dmvnorm(x, mean, covariance)
}
\arguments{
\item{x}{is a n x d matrix of real values}

\item{mean}{is the mean vector}

\item{covariance}{is the covariance matrix}
}
\value{
a vector of n log-density evaluations
}
\description{
Function to evaluate a multivariate Normal density of each row of the given matrix.
}
\details{
This function does not check anything (i.e. that the given
covariance is PSD). Thus it is faster than functions in standard packages, but more risky.
}
\examples{
x <- rmvnorm(10, rep(0, 5), diag(1, 5, 5))
dmvnorm(x, rep(-1, 5), diag(2, 5, 5))
# mvtnorm::dmvnorm(x, rep(-1, 5), diag(2, 5, 5), log = T)
}
\seealso{
\code{\link{rmvnorm}}, \code{\link{rmvnorm_cholesky}}, \code{\link{dmvnorm_cholesky_inverse}}
}
