% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariatenormal.R
\name{dmvnorm_cholesky_inverse}
\alias{dmvnorm_cholesky_inverse}
\title{Evaluate density of multivariate Normal}
\usage{
dmvnorm_cholesky_inverse(x, mean, cholesky_inverse)
}
\arguments{
\item{x}{is a n x d matrix of real values}

\item{mean}{is the mean vector}

\item{cholesky_inverse}{is a Cholesky factor of the inverse of the covariance matrix; for instance obtained
as \code{t(chol(solve(covariance)))}.}
}
\value{
a vector of n log-density evaluations
}
\description{
Function to evaluate a multivariate Normal density of each row of the given matrix,
where the covariance matrix is specified through a Cholesky factor of its inverse.
}
\details{
This function does not check anything (i.e. that the given
covariance is PSD). Thus it is faster than functions in standard packages, but more risky.
}
\examples{
x <- rmvnorm(10, rep(0, 5), diag(1, 5, 5))
dmvnorm_cholesky_inverse(x, rep(-1, 5), t(chol(solve(diag(2, 5, 5)))))
# mvtnorm::dmvnorm(x, rep(-1, 5), diag(2, 5, 5), log = T)
}
\seealso{
\code{\link{rmvnorm}}, \code{\link{rmvnorm_cholesky}}, \code{\link{dmvnorm}}
}
