% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{multinomial_resampling_R}
\alias{multinomial_resampling_R}
\title{Multinomial resampling (slow implementation)}
\usage{
multinomial_resampling_R(nsamples, normalized_weights)
}
\arguments{
\item{nsamples}{is a number of desired samples}

\item{normalized_weights}{is a vector of probabilities (non-negative values summing to one)}
}
\value{
a vector of \code{nsamples} integers between 1 and the size of \code{normalized_weights}.
}
\description{
Sample \code{nsamples} times from a categorical
distribution with probabilities given by \code{normalized_weights}.
}
\examples{
N <- 1000
logweights <- rnorm(N)
normalize_weight_results <- normalize_weight(logweights)
normalized_weights <- normalize_weight_results$nw
multinomial_resampling_R(10, normalized_weights)
}
