% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_weight.R
\name{normalize_weight}
\alias{normalize_weight}
\title{Normalize vector of log-weights}
\usage{
normalize_weight(logweights)
}
\description{
Takes a vector of real values, and return the vector of normalized exponentiated values,
as well as average of the exponentiated but unnormalized values.
}
