% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet.R
\name{pet}
\alias{pet}
\title{A Pet Decision Function
This function helps the user to determine what kind of pet they should get}
\usage{
pet(x)
}
\description{
A Pet Decision Function
This function helps the user to determine what kind of pet they should get
}
