% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{ssp_resampling}
\alias{ssp_resampling}
\title{SSP resampling}
\usage{
ssp_resampling(nsamples, normalized_weights)
}
\arguments{
\item{normalized_weights}{is a vector of probabilities (non-negative values summing to one)}
}
\value{
A vector of ancestors of the same length as \code{normalized_weights}
}
\description{
Srinivasan Sampling Process (SSP) resampling returns a vector of \code{nsamples} ancestors
(integers from 1 to \code{N}) based on a given vector of \code{N} normalized weights.
This function internally uses a vector of \code{N} independent Unif(0,1).

The SSP resampling is unbiased: if \code{A = ssp_resampling(N,W)}, then
for every \code{i = 1,...,N}, the expectation of \code{sum(A==i)} is equal to \code{N*W[i]}.

The SSP resampling has the additional property that \code{sum(A==i)} is exactly equal to either
\code{floor(N*W[i])} or \code{(1+floor(N*W[i]))}.
}
\details{
See Negative association, ordering and convergence of resampling methods, by Gerber,
Chopin, and Whiteley (2017) [https://arxiv.org/abs/1707.01845]
}
\examples{
N <- 1000
logweights <- rnorm(N)
normalize_weight_results <- normalize_weight(logweights)
normalized_weights <- normalize_weight_results$nw
ssp_resampling(normalized_weights)
}
