% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{systematic_resampling}
\alias{systematic_resampling}
\title{Systematic resampling}
\usage{
systematic_resampling(nsamples, normalized_weights)
}
\arguments{
\item{nsamples}{is a number of desired samples}

\item{normalized_weights}{is a vector of probabilities (non-negative values summing to one)}
}
\value{
a vector of \code{nsamples} integers between 1 and the size of \code{normalized_weights}.
}
\description{
Systematic resampling returns a vector of ancestors based on the given
weights, using only one uniform.
}
\details{
See Comparison of Resampling Schemes for Particle Filtering, Douc, Cappe and Moulines,
https://arxiv.org/pdf/cs/0507025.pdf
}
\examples{
N <- 1000
logweights <- rnorm(N)
normalize_weight_results <- normalize_weight(logweights)
normalized_weights <- normalize_weight_results$nw
systematic_resampling(10, normalized_weights)
}
