% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{tree_getpath}
\alias{tree_getpath}
\title{tree_getpath}
\usage{
tree_getpath(tree, n)
}
\arguments{
\item{tree}{is a \code{list} encoding a tree genereated via the function \code{tree_init} or
previously updated via the function \code{tree_update}.}

\item{n}{the index of the path to extract.}
}
\value{
This function outputs a matrix of size \code{dimX} by \code{nstep}
(see documentation of \code{tree_init}) corresponding to the path indexed by \code{n}.
}
\description{
This function extracts a path from a tree.
}
\examples{
N = 5
dimX = 2
nstep = 10

#--- Artificially create N particles, each of dimension dimX, evolved for nstep timesteps ---#
X = array(rnorm(N*dimX*(nstep+1)), dim = c(dimX,N,nstep))
a = replicate(nstep-1, sample(1:N, size = N, replace = TRUE))
tree = tree_init(X[,,1])
for (t in 2:nstep){
tree = tree_update(tree, X[,,t], a[,t-1])
}

#--- List all the N final paths ---#
for (i in 1:N){
cat("Path",toString(i),"\\n")
print(tree_getpath(tree, i))
}
}
