% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{tree_init}
\alias{tree_init}
\title{tree_init}
\usage{
tree_init(X_0)
}
\arguments{
\item{X_0}{is a \code{dimX} by \code{N} matrix representing a set of \code{N} initial
\code{dimX}-dimensional particles.}
}
\value{
This function outputs a \code{list} containing the following objects: \describe{
  \item{\code{N}}{Number of particles (i.e. \code{ncol(X_0)}).}
  \item{\code{M}}{Size of the buffer (initialized at \code{10*N}, increased automatically when needed).}
  \item{\code{dimX}}{Dimension of each particle (i.e. \code{nrow(X_0)}).}
  \item{\code{a_star}}{Vector of length \code{M} encoding ancestors (see details).}
  \item{\code{o_star}}{Vector of length \code{M} encoding number of offsprings (see details).}
  \item{\code{x_star}}{Matrix of size \code{dimX} by \code{M} encoding particles (see details).}
  \item{\code{l_star}}{Vector of length \code{N} containing the indices of the leaves (see details).}
  \item{\code{nstep}}{Height of the tree (starting at 0 with the root nodes).}
}
}
\description{
This function initializes a tree (encoded as a \code{list}).
}
\details{
See \emph{Path storage in the particle filter}, by Jacob, Murray, Rubenthaler (2015)
[\url{https://link.springer.com/article/10.1007/s11222-013-9445-x}]
}
\examples{
N = 5
dimX = 2
X_0 = matrix(rnorm(N*dimX), nrow = dimX, ncol = N)
tree = tree_init(X_0)
}
